Allow anonymizer to retrieve attributes from factories
This allows 2 methods in the factories to be defined. Both return an array.
The first method, anonymizableAttributes(), returns an array of keys that the factory definition already has. The anonymizer will use the factory’s definition when anonymizing. This will reduce the necessary amount of code, as it reuses existing code.
The second method, anonymizableDefinition, is the same as the method that can be defined on the model. It returns a keyed array with predefined values. This can be used to anonymize using custom data instead of relying on the factories definition.

Both method can be used at once